+++ HISTORY.md
@@ -1,5 +1,6 @@
 ## v1.0.1
 * Added Cloudflare Turnstile `TestingSiteKeys` and `TestingSecretKeys`.
+* Changed `ClientResponseKey` -> `ClientResponseFormKey`.
 
 ## v1.0.0
 * Initial release.