+++ HISTORY.md
@@ -1,7 +1,7 @@
 ## v1.0.1
 * Added Cloudflare Turnstile `TestingSiteKeys` and `TestingSecretKeys`.
 * Changed `ClientResponseKey` -> `ClientResponseFormKey`.
-* VerifyAsync method signature changed; made remoteIp param optional per documnetation.
+* VerifyAsync method signature changed; made remoteIp param optional per documentation.
 * Support `CData` field in `TurnstileResponse`.
 
 ## v1.0.0