+++ HISTORY.md
@@ -1,3 +1,6 @@
+## v1.0.2
+* Added support for `action` response value.
+
 ## v1.0.1
 * Added Cloudflare Turnstile `TestingSiteKeys` and `TestingSecretKeys`.
 * Changed `ClientResponseKey` -> `ClientResponseFormKey`.